# Optimized packing pointers from json