In my test I had some 200000 pointers, and with the linear searches and O^2 complexity of the json parser it took too long. I waited ~15 minutes before I aborted the test. After these changes my test have the following timings:
DL Binary write 4.1s
DL Json write 4.6s
DL Json read 2.7s
DL Binary read 1.9s